Improved the Style singleton to support different style configs
The different styleConfigs are loaded using the styleConfig variable
in am-config.yaml. By default it's set to "auto" which will try to adapt
the UI to the current screen size.

Also consolidated all the different Main*.qml files into two versions
Main.qml for displaying the HUD and ClusterAndHUD.qml for showing the
InstrumentCluster in a separate Window together with the HUD.

Change-Id: I8332cb97528db5de4fcb478e4f943ed93174d84f
Task-number: QTAUTO-21
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>