Added a ScreenManager plugin and use it in ClusterAndHUD.qml
Currently there is no QML api for selecting which screen a window
should use, this is a small helper which makes it possible to do that
from QML.

The ClusterAndHUD.qml now moves the cluster to the second Screen. This
only works if the QPA handles the screens separately, which means in a
modern desktop environment the screen will be set for the window, but
it won't move there.

As before the ClusterAndHUD.qml can only be started on system with a
WindowManager running or a system with 2 screens

Change-Id: Ia854dd420d72c8349c90b19971c88a27adab7bdf
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>