Removed popup wayland surface
There was a popup support which was creating a new wayland
surface for each app. With this commit this approach was
removed since we would like to use org.freedesktop.Notification
both for notifications and popups. In that way we would
render popup just on one place and the system ui would
have the ownership.

Change-Id: I4f8eb5faccc4de4eb3b574ea1d6f5ab27cae4dc7
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>