Changed the compositor to handle stopping an application
The UI usually doesn't stop an application, instead just the visibility
is changed and the surface items are not deleted at all.

If an application is stopped (e.g. by stopping it using the
appman-controller) we need to make sure to first run the closing animation
before we actually release the window.

Otherwise the closing animation looks broken and we leave the window
StackView in an broken state

Change-Id: I0d184283d7b0f6b41910435e328d286ed7bb1c76
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>