Improve window handling
When starting an application in the minimized state we still need to
accept the window, but reparent it to a dummy item instead of pushing it
into the window stack. This is achieved by extending the
applicationSurfaceReady signal with a isMinimized property.

Also use the new isClosing property of the application manager to
identify closing windows. This is needed to make sure we only raise
windows which are in a usable state.

Change-Id: I0a5d544576859024b3f33e0814e72707bc692621
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>
Reviewed-by: Erik Botö <erik.boto@pelagicore.com>