Improve window handling
Always maintain a list of all windows accepted by the system-ui
and their type. Depending on the type we know what needs to be
done in case the application is closed or stopped.

This is needed to properly release the surfaces of all windows
we accepted. IVI windows which are not minimized need to poped
from the stack first and a closing animation be played. All other
surfaces can be released once the surfaces are lost.

Change-Id: Ie8ede8c2b14b288d1a182d7c32095c2c4c90b845
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>