Added support for stacking windows in the LaunchController
It's now possible to start multiple applications and stack the windows
on each other. If a window is currently visible and an second application
was started, the new window will be stacked on top. The back button
will reveal the previous window again.

If an window in the stack which is not the current window is closed.
The window is removed from the stack without animations

Change-Id: I0d2b0010ffe4ec07bb075d53c5ba524b52d55a7c
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>