Performance improvement
The ListView of MenuScreen doesn't hide Item's which are currently not
on the screen. The logic has been improved to hide all Items which
are not on the Screen and are not needed for a transition.

In addition we make sure that the surface for the mapWidget is hidden
once the HomeScreen is not visible.

Change-Id: I1974ce997811366fe9f011612b5fa5d91aa7e76d
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>