Ventilation item: fix changes by more than one step
Previously, two things did not work:
- Changing by more than one level by directly tapping on the
  desired level. This was apparently intended to work but
  didn't, from looking at the code.
- Changing by more than one level from the backend

In both cases, the NumberAnimation on the view's index which was
aliased to the ventilation level caused a changed signal for
every step change, which, because it was the most recent change,
became *the* change and caused the final value to fall short of
the actual target value.
After the core fix, a similar fix to the previous one is needed
to make the backend to UI propagation work at all again. But
then it works right.

Change-Id: I31d7f247203cf25b17d06d0eb4edc7af81af4334
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>