Fix the problems with UI not reflecting the QtIvi backend changes
In current implementation the property bindings between QtIvi backend
properties and GUI object properties are used for two-way communication.
In case GUI requests a change, the requested GUI object property is assigned
with a new value. This operation removes property binding to the QtIvi
backend. Additional code in onXXXChanged: handler must be written explicitly
to propagate the change to the backend and recreate the binding afterwards.
This scheme is difficult to maintain and test.

In this modification a second channel is introduced - the property bindings
take care of propagation of updates from backend to GUI, and functions are
used by GUI in order to request backend changes. GUI itself does not update
it's own internal variables on user actions, and instead relies on property
bindings to follow the change in backend - this approach makes the code
simpler, avoids binding recreation, and handles well cases where backend
properties changes value to a different than assigned to it - backend can
react to property assignments in non-trivial ways.

Change-Id: I55423fe806d1db9d099c7983676dd1afeb1b5b69
Reviewed-by: Andreas Hartmetz <andreas.hartmetz@kdab.com>
Reviewed-by: Harri Porten <porten@froglogic.com>
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>