Music app: Fix inconsitency between initial song metadata and song
It is not enough to reevaluate metadata when count changes, since
different queries can have the same number of results. This is why
the initial song if you just launch the app and press play is not
the song you see title and art for. This commit fixes that by
acting on query changes instead.

Change-Id: Ifea5eb4090c8825a4270e98f89b16ad9e3740ec3
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>