Make it easy to use the qml plugins from the build folder
Added a new qmlplugin qmake feature. This contains
all the common code of the qml plugins and makes
sure the plugins end up in within a qml folder
inside the build-folder, which is usable to import
the plugins directly without installation.

In addition a plugins.yaml file is generated, which
contains all the additional settings, to use the
plugins from the build folder.

Change-Id: I1a24236e1132285917d2702f5143a4bc49b3d4e1
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>