Improve Neptune UI documentation style
Updates to the Neptune UI documentation style:
- previous and next page navigation added
- \table and \list commands used when needed
- a new page added for the external page links
- a new TOC page added. It helps the navigation in the
content list in the Qt Creator's help mode.
- a new Getting Started page added. It contains links to the
pages that were earlier under Quick links. Also Neptune UI
components are listed on the page.
- Pelagicore lisence header added in case it was missing
- \caption commands removed under some images. The current font style
of \caption is very close to the normal font style (fixed in online
but still an issue in offline documentation). This is confusing for
the end user.
- sentences simplified according to Qt writing style guide

Task-number: QTAUTO-527
Change-Id: I534e20149eacae3759ab9307449b2235a6f35abd
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>