Support for selecting window switching animations
Added a new "Tumble" animation, in addition to the existing "Zoom" one.
I would like to do some fancier animations but I didn't find out
a way to set the transform property of a target item from within a Transition.

The setting UI for selecting the window switch animation is tucked into
the About pane as it doesn't belong to the actual IVI UI.

Task-number: QTAUTO-521
Change-Id: If8e04f64904bbd5e77173910d8e746f5ebcd2d69
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>