Convert MouseEvents into TouchEvents on unix
The UI was meant to be used with touch interaction, not mouse.
MouseEvents and TouchEvents follow completely different code paths in Qt.
Therefore when interacting with neptune on a linux desktop using a mouse,
which is the standard development setup, we should still be exercising the
touch code path, which is what happens on target hardware (touchscreen
displays).

Otherwise you might enconter touch vs. mouse specific issues only when
trying the software on a touchscreen, and that's not very productive.

Change-Id: Iedcf2052a15539d0c9024b233a20ae6a55e3e06d
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>